+++ packages/components/molecules/f-alert/CHANGELOG.md
@@ -4,6 +4,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+Latest – to be added to the next release
+------------------------------
+*February 8, 2022*
+
+### Added
+- Visual regression tests to component
+
+
 v4.2.0
 ------------------------------
 *February 2, 2022*